dataGatherer/requirements.txt
requests>=2.20.0
python-firebase>=0
